refactor qtmoc
